added final logic to protected routes

+++ index.js
@@ -39,6 +39,9 @@ const verifyToken = async (req, res, next) => {
 
   try {
     const { payload } = await jwtVerify(token, JWKS);
+
+    req.user = payload;
+
     console.log(payload);
     next();
   } catch (error) {
@@ -72,8 +75,19 @@ async function run() {
       res.json(result);
     });
 
-    app.patch("/rooms/:id", async (req, res) => {
+    app.patch("/rooms/:id", verifyToken, async (req, res) => {
       const { id } = req.params;
+
+      const room = await roomCollection.findOne({
+        _id: new ObjectId(id),
+      });
+
+      if (room.ownerEmail !== req.user.email) {
+        return res.status(403).json({
+          message: "Forbidden Access",
+        });
+      }
+
       const updatedData = req.body;
       const result = await roomCollection.updateOne(
         { _id: new ObjectId(id) },
@@ -85,50 +99,109 @@ async function run() {
 
     app.delete("/rooms/:id", verifyToken, async (req, res) => {
       const { id } = req.params;
+
+      const room = await roomCollection.findOne({
+        _id: new ObjectId(id),
+      });
+
+      if (room.ownerEmail !== req.user.email) {
+        return res.status(403).json({
+          message: "Forbidden Access",
+        });
+      }
+
       const result = await roomCollection.deleteOne({ _id: new ObjectId(id) });
 
       res.json(result);
     });
 
-    app.post("/rooms", async (req, res) => {
+    app.post("/rooms", verifyToken, async (req, res) => {
       const roomData = req.body;
       // console.log(roomData);
+
+      roomData.ownerEmail = req.user.email;
+
+      roomData.bookingCount = 0;
 
       const result = await roomCollection.insertOne(roomData);
 
       res.json(result);
     });
 
-    app.post("/my-bookings", async (req, res) => {
-      const bookingData = req.body;
-      const { roomId, date, startTime, endTime } = bookingData;
+    app.get("/my-listings/:email", verifyToken, async (req, res) => {
+      const email = req.params.email;
 
+      const query = {
+        ownerEmail: email,
+      };
+
+      const result = await roomCollection.find(query).toArray();
+
+      res.send(result);
+    });
+
+    app.post("/my-bookings", verifyToken, async (req, res) => {
       try {
+        const bookingData = req.body;
+
+        const { roomId, date, startTime, endTime } = bookingData;
+
         const conflict = await bookingsCollection.findOne({
           roomId: roomId,
           date: date,
           status: "confirmed",
-          startTime: { $lt: endTime },
-          endTime: { $gt: startTime },
+
+          startTime: {
+            $lt: endTime,
+          },
+
+          endTime: {
+            $gt: startTime,
+          },
         });
+
         if (conflict) {
           return res.status(400).json({
             success: false,
             message: "This room is already booked for the selected time slot.",
           });
         }
-        const result = await bookingsCollection.insertOne({
-          ...bookingData,
-          status: "confirmed",
-        });
 
-        res.status(201).json({ success: true, ...result });
+        bookingData.userEmail = req.user.email;
+
+        bookingData.status = "confirmed";
+
+        bookingData.createdAt = new Date();
+
+        const result = await bookingsCollection.insertOne(bookingData);
+
+        await roomCollection.updateOne(
+          {
+            _id: new ObjectId(roomId),
+          },
+          {
+            $inc: {
+              bookingCount: 1,
+            },
+          }
+        );
+
+        res.status(201).json({
+          success: true,
+          message: "Room booked successfully",
+          insertedId: result.insertedId,
+        });
       } catch (error) {
-        res.status(500).json({ success: false, message: error.message });
+        console.log(error);
+
+        res.status(500).json({
+          success: false,
+          message: "Failed to book room",
+        });
       }
     });
 
-    app.get("/my-bookings/:userId", async (req, res) => {
+    app.get("/my-bookings/:userId", verifyToken, async (req, res) => {
       const { userId } = req.params;
 
       const result = await bookingsCollection
@@ -140,6 +213,16 @@ async function run() {
 
     app.delete("/my-bookings/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+
+      const booking = await bookingsCollection.findOne({
+        _id: new ObjectId(id),
+      });
+
+      if (booking.userEmail !== req.user.email) {
+        return res.status(403).json({
+          message: "Forbidden Access",
+        });
+      }
 
       const query = { _id: new ObjectId(id) };
